short -> int in setCount
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@142 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/Header.java
@@ -97,8 +97,8 @@ getOpcode() {
 }
 
 public void
-setCount(int field, short value) {
-	counts[field] = value;
+setCount(int field, int value) {
+	counts[field] = (short) value;
 }
 
 public void
